Экран "Медиатека", вынес адаптер в отдельный класс.
