+++ app/src/main/java/com/example/playlistmaker/mediateka/ui/MediaActivity.kt
@@ -5,9 +5,6 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
-import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.example.playlistmaker.R
 import com.example.playlistmaker.databinding.ActivityMediatekaBinding
 import com.example.playlistmaker.mediateka.ui.favorites.FavoritesFragment
@@ -50,15 +47,5 @@ class MediaActivity : AppCompatActivity() {
         TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
             tab.text = fragments[position].second
         }.attach()
-    }
-
-    private class MediaPagerAdapter(
-        fragmentActivity: FragmentActivity,
-        private val fragments: List<Fragment>
-    ) : FragmentStateAdapter(fragmentActivity) {
-
-        override fun getItemCount(): Int = fragments.size
-
-        override fun createFragment(position: Int): Fragment = fragments[position]
     }
 }
