+++ app/src/main/java/com/example/playlistmaker/mediateka/ui/MediaPagerAdapter.kt
@@ -1,0 +1,15 @@
+package com.example.playlistmaker.mediateka.ui
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
+import androidx.viewpager2.adapter.FragmentStateAdapter
+
+class MediaPagerAdapter(
+    fragmentActivity: FragmentActivity,
+    private val fragments: List<Fragment>
+) : FragmentStateAdapter(fragmentActivity) {
+
+    override fun getItemCount(): Int = fragments.size
+
+    override fun createFragment(position: Int): Fragment = fragments[position]
+}
